Run findbugs

+++ src/main/java/moze_intel/projecte/gameObjs/tiles/RelayMK1Tile.java
@@ -117,7 +117,7 @@ public class RelayMK1Tile extends TileEmcProducer implements IInventory, ISidedI
 		}
 		else 
 		{
-			this.sendEmcToRequesting(chargeRate / numRequesting);
+			this.sendEmcToRequesting(chargeRate / (double) numRequesting);
 			this.removeEmc(chargeRate);
 		}
 	}
